+++ tests/testthat/test-Report_OverviewText.R
@@ -11,6 +11,7 @@ test_that("It handles different lSetup group values correctly", {
   )
 
   lSetup$GroupLevel <- "Country"
+  lStudy$CountryCount <- "3"
   expect_output(
     Report_OverviewText(lSetup, reportingResults, lStudy),
     "countries"
@@ -25,9 +26,9 @@ test_that("Data filtering checks", {
   )
   lStudy <- list(ParticipantCount = "1301", SiteCount = "176")
 
-  expect_output(Report_OverviewText(lSetup, reportingResults, lStudy), "29 sites have at least one red KRI")
-  expect_output(Report_OverviewText(lSetup, reportingResults, lStudy), "132 sites have at least one red or amber KRI")
-  expect_output(Report_OverviewText(lSetup, reportingResults, lStudy), "158 sites have neither red nor amber KRIS and are not shown")
+  expect_output(Report_OverviewText(lSetup, reportingResults, lStudy), "sites have at least one red KRI")
+  expect_output(Report_OverviewText(lSetup, reportingResults, lStudy), "sites have at least one red or amber KRI")
+  expect_output(Report_OverviewText(lSetup, reportingResults, lStudy), "sites have neither red nor amber KRIS and are not shown")
 })
 
 test_that("Handles empty dataframe cases", {
@@ -63,5 +64,5 @@ test_that("Handles different flag configurations", {
 
   expect_output(Report_OverviewText(lSetup, dfSummary, lStudy), "2 sites have at least one red KRI")
   expect_output(Report_OverviewText(lSetup, dfSummary, lStudy), "3 sites have at least one red or amber KRI")
-  expect_output(Report_OverviewText(lSetup, dfSummary, lStudy), "1 sites have neither red nor amber KRIS and are not shown")
+  expect_output(Report_OverviewText(lSetup, dfSummary, lStudy), "0 sites have neither red nor amber KRIS and are not shown")
 })