+++ tests/testthat/test-util-MakeMetricTable.R
@@ -4,8 +4,10 @@ test_that("Empty dfs return empty dfs", {
   expect_equal(
     MakeMetricTable(dfResults_empty, dfGroups_empty),
     data.frame(
-      Group = character(), Enrolled = character(), Numerator = double(),
-      Denominator = double(), Metric = double(), Score = double(),
+      StudyID = character(), GroupID = character(), MetricID = character(),
+      Group = character(), SnapshoteDate = as.Date(integer()),
+      Enrolled = integer(), Numerator = integer(),
+      Denominator = integer(), Metric = double(), Score = double(),
       Flag = character()
     )
   )
@@ -18,7 +20,10 @@ test_that("Correct data structure when proper dataframe is passed", {
   expect_s3_class(result, "data.frame")
   expect_setequal(
     colnames(result),
-    c("Group", "Enrolled", "Numerator", "Denominator", "Metric", "Score", "Flag")
+    c(
+      "StudyID", "GroupID", "MetricID", "Group", "SnapshotDate", "Enrolled",
+      "Numerator", "Denominator", "Metric", "Score", "Flag"
+    )
   )
 })
 
@@ -42,7 +47,7 @@ test_that("Score rounding works correctly", {
   reportingResults_filt <- reportingResults %>%
     dplyr::filter(MetricID == unique(reportingResults$MetricID)[[1]])
   result <- MakeMetricTable(reportingResults_filt, reportingGroups)
-  expect_true(any(grepl("3.02", as.character(result$Score))))
+  expect_true(any(grepl("^\\d+\\.\\d{2}$", as.character(result$Score))))
 })
 
 test_that("Errors informatively when multiple MetricIDs passed in", {
@@ -50,4 +55,41 @@ test_that("Errors informatively when multiple MetricIDs passed in", {
     MakeMetricTable(reportingResults, reportingGroups),
     class = "gsm_error-multiple_values"
   )
+})
+
+test_that("Enrolled is an integer", {
+  reportingResults_filt <- reportingResults %>%
+    dplyr::filter(MetricID == unique(reportingResults$MetricID)[[1]])
+  result <- MakeMetricTable(reportingResults_filt, reportingGroups)
+  expect_type(result$Enrolled, "integer")
+})
+
+test_that("Output is expected object", {
+  zero_flags <- c("0X085", "0X086")
+  flags <- c("0X052", "0X027", "0X166")
+
+  reportingResults_filt <- reportingResults %>%
+    FilterByLatestSnapshotDate() %>%
+    dplyr::filter(
+      MetricID == unique(reportingResults$MetricID)[[1]],
+      GroupID %in% c(zero_flags, flags)
+    ) %>%
+    # Add an NA row back for representation.
+    dplyr::bind_rows(
+      tibble::tibble(
+        GroupID = "0X000",
+        GroupLevel = "Site",
+        Numerator = 4L,
+        Denominator = 8L,
+        Metric = 0.5,
+        Score = NA,
+        Flag = NA,
+        MetricID = "Analysis_kri0001",
+        SnapshotDate = as.Date("2012-12-31"),
+        StudyID = "AA-AA-000-0000"
+      )
+    )
+  expect_snapshot({
+    MakeMetricTable(reportingResults_filt, reportingGroups)
+  })
 })