+++ tests/testthat/testdata/create_double_data.R
@@ -1,6 +1,6 @@
 # yaml workflow
 test_wf <- MakeWorkflowList(
-  strNames = "kri0001"
+  strNames = "kri0001", strPath = "workflow/2_metrics"
 )
 test_mapping <- MakeWorkflowList(
   strPath = test_path("testdata/mappings"),
@@ -12,8 +12,8 @@ lRaw <- UseClindata(
     "Raw_AE" = "clindata::rawplus_ae"
   )
 )
-lMapped <- quiet_RunWorkflows(lWorkflow = test_mapping, lData = lRaw)
-lResults <- quiet_RunWorkflows(lWorkflow = test_wf, lData = lMapped)
+lMapped <- quiet_RunWorkflows(lWorkflows = test_mapping, lData = lRaw)
+lResults <- quiet_RunWorkflows(lWorkflows = test_wf, lData = lMapped)
 
 # functional workflow
 Mapped_SUBJ <- clindata::rawplus_dm %>%